requirements.txt
numpy
pandas<0.25.0
opencv-python-headless
pyyaml
coloredlogs
scikit-learn
scipy
matplotlib
peewee
pytest